Begin process of hiding GPIO implementation details
We have a lot of branching going on with controlling the IO on the
board. I want to dispatch instead on something like a class so we can
more easily factor out common code and create an interface that we could
have multiple implementations against. Part of this is also getting rid
of some globals so we no longer let the pin_rain_sense global escape the
module. This means we need to pull check_rain inside of the pins module
which is fine