Finish creating interface classes
Now we have a common class that handles the pins when a shift register
is involved for the RPi and Beagle Bone when using a shift register. I
removed references to disableShiftRegisterOutput that was commented out
in other modules to avoid confusion in the future now that those
functions are hidden. It's okay to hide them since they weren't used
anywhere. I also condensed the setup code and removed some redundancy