Split out pin interfaces into separate modules
Most of the work in this commit is to sort out some of the circular
dependency problems and import-time logic and turning into non-import
logic. I created a new security module that holds functions that were in
helpers that can be safely imported by anything at any time. I moved
some functions from helpers into gv which was the only module calling
it. Finally I updated some of the exception and logging logic to let me
know when things go wrong